Quick POC of a shadowing classloader
Make bridged proxy objects without invoking superclass constructors:
where constructors have side-effects, this is pretty critical. The
CGLIB proxy objects are instances of a subclass of the proxied class;
however, when we make a proxy object, we don't want the base class'
constructor to be executed.

Use Objenesis to make these instances instead.